requirements.txt
asgiref==3.6.0
attrs==22.1.0
black==22.12.0
click==8.1.3
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
iniconfig==1.1.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.11.0
platformdirs==2.6.2
pluggy==1.0.0
pycodestyle==2.9.1
pyparsing==3.0.9
pytz==2022.7.1
sqlparse==0.4.3
